mgyp assignment
