+++ tools/Assembly/assign_MGYP/new_config
@@ -1,6 +1,5 @@
 {
     "mapping"     : "/nfs/production/interpro/metagenomics/peptide_db/mapping",
-    "releases"    : "/nfs/production/interpro/metagenomics/peptide_db/releases",
     "max_pep_length"    : 10000,
     "peptides"    : "/nfs/production/interpro/metagenomics/assembly_peptides_v5"
 }
