+++ tools/Assembly/assign_MGYP/test.fasta
@@ -25,3 +25,9 @@ CRWQRAPRLAPHGAFFARYFGELPALIRLVARSGFNFLRIRKVVARFERHFAEHYARWS
 VLCLPLVTQRAAHWRTRRPLRRYATPSPFWKRLQGISQLFAIRGAELGEGSTPT
 >ERZ1292630.716681-Hain-H41-02um-R3-716681-length-500-cov-0_3_251_-
 MLASQGLSSAQEGGARADYRIVAHLDGATKRLTGDLELRWTNASGEAVEELWFHLYLNAFSNNRSSHLVESQGKLRSKEMSAG
+>ERZ1292630.1-Hain-H41-02um-R3-1-length-778660-cov-10_2 # 1075 # 1335 # -1 # ID=1_2;partial=00;start_type=ATG;rbs_motif=None;rbs_spacer=None;gc_cont=0.444
+MFIFLIGSISATDIGLAKNSESAILIEASTGKILFEKEKDKKKALISNEGNVEISPNDVK
+NKAFFWIFMIGFFGMGLVCSAPMRQMNPYISDIYGSEFAASVIGLSSLMGIFGKILLGAL
+HDKFGSIKSSSIAFVAFALAFIAAIAGAHGEKNLFYVYIILYSFAAGVGTVSAPLIISAT
+FGTKSFNMMRGLTQSPLQAGMSLGGLMVAGVFDITGAYTLGWAACIAISVISIICFYAAH
+KMSRKYAAEINMAK
